requirements.txt
Django==2.1.3
django-webpack-loader==0.6.0
djangorestframework==3.9.0
pytz==2018.7
